started replacing event manager with executor
still need memory allocator for events + guest side support

+++ backends/aws-lambda/iomod/guest/src/lib.rs
@@ -16,7 +16,7 @@ pub mod database {
 
         match event_id {
             -1 => None,
-            _ => Some(Event::new(event_id as usize))
+            _ => Some(Event::new(event_id as u32))
         }
     }
 }
